# Conflicts:
#	src/main/kotlin/dnd11th/blooming/domain/repository/myplant/MyPlantQueryDslRepository.kt
#	src/main/kotlin/dnd11th/blooming/domain/repository/myplant/MyPlantQueryDslRepositoryImpl.kt

+++ src/main/kotlin/dnd11th/blooming/api/dto/user/TokenResponse.kt
@@ -10,9 +10,9 @@ data class TokenResponse(
     @field:Schema(description = "AccessToken", example = "accessToken")
     val accessToken: String,
     @field:Schema(description = "accessToken 만료일자(초)", example = "604800(1주일)")
-    val expiresIn: Int,
+    val expiresIn: Long,
     @field:Schema(description = "RefreshToken", example = "refreshToken")
     val refreshToken: String,
     @field:Schema(description = "refreshToken 만료일자(초)", example = "15552000(6개월)")
-    val refreshTokenExpiresIn: Int,
+    val refreshTokenExpiresIn: Long,
 )
